+++ src/main/java/club/runningcoin/service/VoteRecordService.java
@@ -35,12 +35,12 @@ public interface VoteRecordService {
      * @param id the id of the entity
      * @return the entity
      */
-    Optional<VoteRecordDTO> findOne(Long id);
+    Optional<VoteRecordDTO> findOne(Integer id);
 
     /**
      * Delete the "id" voteRecord.
      *
      * @param id the id of the entity
      */
-    void delete(Long id);
+    void delete(Integer id);
 }
